+++ CHANGES.rst
@@ -1,3 +1,9 @@
+v0.0.4
+######
+- Include examples in documentation
+- Implement :py:class:`Recording <pupil_labs.realtime_api.models.Recording>` model class
+- Add :py:attr:`Status.recording <pupil_labs.realtime_api.models.Status.recording>` attribute
+
 v0.0.3
 ######
 - Move Control.Error to dedicated :py:exc:`ControlError <pupil_labs.realtime_api.control.ControlError>` class